+++ Application/Features/Product/Commands/DeleteProduct/DeleteProductRequestHandler.cs
@@ -1,5 +1,7 @@
 ﻿using Application.Contracts.Persistence;
+using Ardalis.GuardClauses;
 using MediatR;
+using Microsoft.AspNetCore.Http.HttpResults;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -12,7 +14,8 @@ namespace Application.Features.Product.Commands.DeleteProduct
     {
         public async Task Handle(DeleteProductRequest request, CancellationToken cancellationToken)
         {
-            var product = await productRepository.GetByIdAsync(request.ProductId);
+            var product = await productRepository.GetByIdAsync(request.Id);
+            if (product is null) throw new NotFoundException(request.Id.ToString(), "Product");
             await productRepository.DeleteAsync(product);
         }
     }
